Extract LaunchCommand to its own file (detekt TooManyFunctions)
